fixed for migration to raspberry

+++ app/src/main/java/com/example/skolapp_client_gab/VolleyCallBack.java
@@ -1,5 +1,5 @@
 package com.example.skolapp_client_gab;
 
 public interface VolleyCallBack{
-    void onSuccess(float value);
+    void onSuccess(double value);
 }
